+++ inc/filters.php
@@ -34,5 +34,25 @@ add_filter( 'wp_kses_allowed_html', function( $array ) {
 		'id'    => array(),
 	);
 
+	$array['strong'] = array(
+		'class' => array(),
+		'id'    => array(),
+	);
+
+	$array['b'] = array(
+		'class' => array(),
+		'id'    => array(),
+	);
+
+	$array['i'] = array(
+		'class' => array(),
+		'id'    => array(),
+	);
+
+	$array['em'] = array(
+		'class' => array(),
+		'id'    => array(),
+	);
+
 	return $array;
 }, 44 );